+++ src/icemac/ab/calendar/browser/renderer/tests/test_base.py
@@ -2,6 +2,7 @@
 from icemac.ab.calendar.browser.renderer.base import Calendar
 from icemac.ab.calendar.browser.renderer.interfaces import IRenderer
 from zope.interface.verify import verifyObject
+import pytest
 
 
 def test_base__Calendar__1():
@@ -14,3 +15,22 @@ def test_base__Calendar__write__1():
     calendar = Calendar(None, None, None)
     calendar.write(u'aäöü%s', u'ßen')
     assert u'aäöüßen\n' == calendar.read()
+
+
+def test_base__Calendar__update__1():
+    """It does nothing.
+
+    Test is only here to complete test coverage.
+    """
+    calendar = Calendar(None, None, None)
+    calendar.update()
+
+
+def test_base__Calendar__render__1():
+    """It raises an exception: method has to be implemented by child classes.
+
+    Test is only here to complete test coverage.
+    """
+    calendar = Calendar(None, None, None)
+    with pytest.raises(NotImplementedError):
+        calendar.render()